event: Move read_events_file to utils/event.c
read_events_file() is better to be placed in utils/event.c rather than
utils/data-file.c so this patch moves the function and slightly modify
its comment.
