+++ utils/event.h
@@ -35,4 +35,6 @@ struct uftrace_pmu_branch {
 char * event_get_name(struct uftrace_data *handle, unsigned evt_id);
 char * event_get_data_str(unsigned evt_id, void *data, bool verbose);
 
+int read_events_file(struct uftrace_data *handle);
+
 #endif  /* UFTRACE_EVENT_H */
